removed lifedyclestore from dockercontainerlogic base class
