+++ common/src/main/java/de/uniulm/omi/cloudiator/lance/application/component/DeployableComponent.java
@@ -1,8 +1,10 @@
 package de.uniulm.omi.cloudiator.lance.application.component;
 
 import de.uniulm.omi.cloudiator.lance.lca.container.ContainerException;
+import de.uniulm.omi.cloudiator.lance.lifecycle.LifecycleStore;
 
 public class DeployableComponent extends AbstractComponent {
+  private final LifecycleStore lifecycle;
 
   @Override
   public boolean isDynamicComponent() {
@@ -25,29 +27,40 @@ public class DeployableComponent extends AbstractComponent {
   }
 
   public static class Builder extends AbstractComponent.Builder<Builder> {
+    private volatile LifecycleStore lifecycle;
 
-      public Builder(String name, ComponentId id) {
-        //this.lifecycle = lifecycle;
-        this.nameParam = name;
-        this.myIdParam = id;
-      }
+    public Builder(String name, ComponentId id) {
+      //this.lifecycle = lifecycle;
+      this.nameParam = name;
+      this.myIdParam = id;
+    }
 
-      public static Builder createBuilder(String name, ComponentId componentId) {
-        return new Builder(name,  componentId);
-      }
+    public Builder addLifecycleStore(LifecycleStore lifecycleStore) {
+      this.lifecycle = lifecycleStore;
+      return this;
+    }
 
-      @Override
-      public DeployableComponent build() {
-        return new DeployableComponent(this);
-      }
+    public static Builder createBuilder(String name, ComponentId componentId) {
+      return new Builder(name,  componentId);
+    }
 
-      @Override
-      protected Builder self() {
+    @Override
+    public DeployableComponent build() {
+                                       return new DeployableComponent(this);
+                                                                             }
+
+    @Override
+    protected Builder self() {
         return this;
       }
-    }
+  }
 
-    private DeployableComponent(Builder builder) {
-        super(builder);
-    }
+  private DeployableComponent(Builder builder) {
+      super(builder);
+      lifecycle = builder.lifecycle;
+  }
+
+  public LifecycleStore getLifecycleStore() {
+    return lifecycle;
+  }
 }
